avg_mark: int>float teacher>isTeacher

+++ src/main/java/ru/cft/shift/repetito/entity/response/UserSimpleResponse.java
@@ -19,7 +19,7 @@ public class UserSimpleResponse {
     private String lastName;
     private String paronym;
     private int price;
-    private int avgMark;
+    private float avgMark;
 
     public Long getId() {
         return id;
@@ -61,11 +61,11 @@ public class UserSimpleResponse {
         this.price = price;
     }
 
-    public int getAvgMark() {
+    public float getAvgMark() {
         return avgMark;
     }
 
-    public void setAvgMark(int avgMark) {
+    public void setAvgMark(float avgMark) {
         this.avgMark = avgMark;
     }
 }
